fix: move searchNpgOperations service logic from helpdesk to ecommerce
