+++ src/main/kotlin/it/pagopa/ecommerce/helpdesk/services/v1/HelpdeskService.kt
@@ -1,22 +1,15 @@
 package it.pagopa.ecommerce.helpdesk.services.v1
 
-import it.pagopa.ecommerce.commons.client.NpgClient
-import it.pagopa.ecommerce.commons.client.NpgClient.PaymentMethod
-import it.pagopa.ecommerce.commons.domain.TransactionId
-import it.pagopa.ecommerce.commons.exceptions.NpgResponseException
-import it.pagopa.ecommerce.commons.generated.npg.v1.dto.OrderResponseDto
 import it.pagopa.ecommerce.commons.utils.ConfidentialDataManager
-import it.pagopa.ecommerce.commons.utils.NpgApiKeyConfiguration
 import it.pagopa.ecommerce.helpdesk.dataproviders.v1.mongo.EcommerceTransactionDataProvider
 import it.pagopa.ecommerce.helpdesk.dataproviders.v1.oracle.PMTransactionDataProvider
-import it.pagopa.ecommerce.helpdesk.exceptions.*
 import it.pagopa.ecommerce.helpdesk.exceptions.InvalidSearchCriteriaException
 import it.pagopa.ecommerce.helpdesk.exceptions.NoResultFoundException
 import it.pagopa.ecommerce.helpdesk.utils.ConfidentialMailUtils
 import it.pagopa.ecommerce.helpdesk.utils.v1.SearchParamDecoder
 import it.pagopa.ecommerce.helpdesk.utils.v1.buildTransactionSearchResponse
-import it.pagopa.generated.ecommerce.helpdesk.model.*
-import java.util.UUID
+import it.pagopa.generated.ecommerce.helpdesk.model.HelpDeskSearchTransactionRequestDto
+import it.pagopa.generated.ecommerce.helpdesk.model.SearchTransactionResponseDto
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
@@ -27,9 +20,7 @@ import reactor.core.publisher.Mono
 class HelpdeskService(
     @Autowired val ecommerceTransactionDataProvider: EcommerceTransactionDataProvider,
     @Autowired val pmTransactionDataProvider: PMTransactionDataProvider,
-    @Autowired val confidentialDataManager: ConfidentialDataManager,
-    @Autowired val npgClient: NpgClient,
-    @Autowired val npgApiKeyConfiguration: NpgApiKeyConfiguration
+    @Autowired val confidentialDataManager: ConfidentialDataManager
 ) {
 
     private val logger = LoggerFactory.getLogger(javaClass)
@@ -174,127 +165,6 @@ class HelpdeskService(
     }
 
     /**
-     * Used by the API that searches for NPG operations. Starting from the transaction id, first it
-     * retrieves the transaction details calling the
-     * [EcommerceTransactionDataProvider.retrieveTransactionDetails] that returns an object of the
-     * data class [NTuple4], then it uses the details to call the [performGetOrderNPG]. Returns a
-     * [SearchNpgOperationsResponseDto], representing a subset of [OrderResponseDto] data.
-     */
-    fun searchNpgOperations(transactionId: String): Mono<SearchNpgOperationsResponseDto> {
-        return ecommerceTransactionDataProvider
-            .retrieveTransactionDetails(transactionId)
-            .flatMap { details ->
-                performGetOrderNPG(
-                    transactionId = TransactionId(transactionId),
-                    orderId = details.first,
-                    pspId = details.second,
-                    correlationId = details.third,
-                    paymentMethod = details.fourth
-                )
-            }
-            .doOnNext { order ->
-                logger.info(
-                    "Performed get order for transaction with id: [{}], last operation result: [{}], operations: [{}]",
-                    transactionId,
-                    order.orderStatus?.lastOperationType,
-                    order.operations?.joinToString { "${it.operationType}-${it.operationResult}" },
-                )
-            }
-            .map { orderResponse ->
-                SearchNpgOperationsResponseDto().apply {
-                    operations =
-                        orderResponse.operations
-                            ?.map { operation ->
-                                OperationDto().apply {
-                                    // Extract authorizationCode and rrn if they exist
-                                    logger.debug(
-                                        "NPG operation additionalData: {}",
-                                        operation.additionalData
-                                    )
-                                    additionalData =
-                                        OperationAdditionalDataDto().apply {
-                                            authorizationCode =
-                                                operation.additionalData
-                                                    ?.get("authorizationCode")
-                                                    ?.toString()
-                                            rrn = operation.additionalData?.get("rrn")?.toString()
-                                        }
-
-                                    operationAmount = operation.operationAmount
-                                    operationCurrency = operation.operationCurrency
-                                    operationId = operation.operationId
-                                    operationResult =
-                                        OperationResultDto.valueOf(
-                                            operation.operationResult.toString()
-                                        )
-                                    operationTime = operation.operationTime
-                                    operationType =
-                                        OperationTypeDto.valueOf(operation.operationType.toString())
-                                    orderId = operation.orderId
-                                    paymentCircuit = operation.paymentCircuit
-                                    paymentEndToEndId = operation.paymentEndToEndId
-                                    paymentMethod =
-                                        operation.paymentMethod?.let {
-                                            PaymentMethodDto.valueOf(it.toString())
-                                        }
-                                }
-                            }
-                            ?.toList()
-                }
-            }
-            .doOnNext { orderResponse ->
-                logger.info(
-                    "Retrieved NPG operations for transaction [{}]: found {} operations",
-                    transactionId,
-                    orderResponse.operations?.size ?: 0
-                )
-            }
-    }
-
-    /**
-     * Imported from transactions-scheduler.
-     *
-     * @see <a
-     *   href="https://github.com/pagopa/pagopa-ecommerce-transactions-scheduler-service/blob/b9ed99fbf6724ba04e646b7b490a43c69c453ebc/src/main/kotlin/it/pagopa/ecommerce/transactions/scheduler/services/TransactionInfoService.kt#L248">performGetOrderNP
-     *   on transactions-scheduler</a>
-     */
-    private fun performGetOrderNPG(
-        transactionId: TransactionId,
-        orderId: String,
-        pspId: String,
-        correlationId: String,
-        paymentMethod: PaymentMethod
-    ): Mono<OrderResponseDto> {
-        logger.info(
-            "Performing get order for transaction with id: [{}], orderId [{}], pspId: [{}], correlationId: [{}], paymentMethod: [{}]",
-            transactionId.value(),
-            orderId,
-            pspId,
-            correlationId,
-            paymentMethod.serviceName,
-        )
-        return npgApiKeyConfiguration[paymentMethod, pspId].fold(
-            { ex -> Mono.error(ex) },
-            { apiKey ->
-                npgClient.getOrder(UUID.fromString(correlationId), apiKey, orderId).onErrorMap(
-                    NpgResponseException::class.java
-                ) { exception: NpgResponseException ->
-                    val responseStatusCode = exception.statusCode
-                    responseStatusCode
-                        .map {
-                            if (it.is5xxServerError) {
-                                NpgBadGatewayException("$it")
-                            } else {
-                                NpgBadRequestException(transactionId.value(), "$it")
-                            }
-                        }
-                        .orElse(exception)
-                }
-            }
-        )
-    }
-
-    /**
      * Calculate pages for display all records given a page size. Return a Pair<Int,Int> where first
      * argument is page size, second one is total count /page size remainder
      */
@@ -308,5 +178,3 @@ class HelpdeskService(
         }
     }
 }
-
-data class NTuple4<T1, T2, T3, T4>(val first: T1, val second: T2, val third: T3, val fourth: T4)
